+++ docs/change-log.md
@@ -3,5 +3,8 @@ outline: deep
 ---
 # Change Log
 
+## 2024-03-15
+* Added pseudo props to Records so that in addition to `user.set('admin', true)` you can also do `user.admin = true` for both setters and getters.
+
 ## 2024-03-12
 * Fixed a security bug where passwords created via the CLI were stored incorrectly.